analysis-service/requirements.txt
streamlit
pandas
psycopg2-binary
altair
google-generativeai>=0.7.2